+++ src/bindings/lua/flux-lua.c
@@ -40,7 +40,7 @@
 #include "src/common/libutil/jsonutil.h"
 #include "src/modules/libmrpc/mrpc.h"
 #include "src/modules/libkz/kz.h"
-#include "src/modules/libzio/zio.h"
+#include "src/common/libsubprocess/zio.h"
 
 #include "json-lua.h"
 #include "kvs-lua.h"
