+++ src/common/Makefile.am
@@ -1,5 +1,5 @@
 SUBDIRS = libtap libev libpmi-client libpmi-server \
-	  liblsd libutil libflux libcompat
+	  liblsd libutil libflux libsubprocess libcompat
 
 AM_CFLAGS = @GCCWARN@ $(CODE_COVERAGE_CFLAGS)
 AM_LDFLAGS = $(CODE_COVERAGE_LDFLAGS)
@@ -15,6 +15,7 @@ libflux_internal_la_LIBADD = \
 	$(builddir)/libev/libev.la \
 	$(builddir)/libpmi-client/libpmi-client.la \
 	$(builddir)/libpmi-server/libpmi-server.la \
+	$(builddir)/libsubprocess/libsubprocess.la \
 	$(builddir)/libcompat/libcompat.la \
 	$(LIBMUNGE) $(JSON_LIBS) $(ZMQ_LIBS) $(LIBPTHREAD) $(LIBUTIL) \
 	$(LIBDL) -lrt
