+++ src/modules/Makefile.am
@@ -1,5 +1,5 @@
 #This order is *important* 
-SUBDIRS = connector-local kvs libmrpc libzio libsubprocess libkz modctl live mecho barrier wreck libjsc resource-hwloc
+SUBDIRS = connector-local kvs libmrpc libkz modctl live mecho barrier wreck libjsc resource-hwloc
 if HAVE_PYTHON
  SUBDIRS+=pymod
 endif
