--- src/modules/libzio/.gitignore
@@ -1,2 +1,0 @@
-/zio-test
-/flux-zio-test
