--- src/modules/libzio/Makefile.am
@@ -1,17 +1,0 @@
-AM_CFLAGS = \
-	@GCCWARN@ \
-	$(CODE_COVERAGE_CFLAGS)
-
-AM_LDFLAGS = \
-	$(CODE_COVERAGE_LDFLAGS)
-
-AM_CPPFLAGS = \
-	$(JSON_CFLAGS) $(ZMQ_CFLAGS) \
-	-I$(top_srcdir) -I$(top_srcdir)/src/include
-
-noinst_LTLIBRARIES = \
-	libzio.la
-
-libzio_la_SOURCES = \
-	zio.c \
-	zio.h
